Fix multi argument filtering behavior for add and edit modes
Add mode does not use filters, so all contexts and projects
from the command line are part of the todo description.

All other edit modes use context and projects mentioned
before any description are used for filtering. It is also
possible to explicitly end the filter list by using the common
double dash marker.

Eg: ttdl edit @first -- @second

The @first context is used for filtering and the @second context
is part of the description.

Signed-off-by: Martin Sivak <mars@montik.net>